+++ rest-openehr/src/main/java/org/ehrbase/rest/openehr/OpenehrDirectoryController.java
@@ -21,7 +21,6 @@ import static org.apache.commons.lang3.StringUtils.unwrap;
 
 import com.nedap.archie.rm.directory.Folder;
 import com.nedap.archie.rm.support.identification.ObjectVersionId;
-import java.net.URI;
 import java.nio.file.InvalidPathException;
 import java.nio.file.Paths;
 import java.time.OffsetDateTime;
@@ -65,7 +64,7 @@ import org.springframework.web.bind.annotation.RestController;
  */
 @TenantAware
 @RestController
-@RequestMapping(path = "${openehr-api.context-path:/rest/openehr}/v1/ehr")
+@RequestMapping(path = BaseController.API_CONTEXT_PATH_WITH_VERSION + "/ehr")
 public class OpenehrDirectoryController extends BaseController implements DirectoryApiSpecification {
 
     private final DirectoryService folderService;
@@ -234,8 +233,7 @@ public class OpenehrDirectoryController extends BaseController implements Direct
             String versionUid = folderDto.getUid().toString();
 
             headers.setETag("\"" + versionUid + "\"");
-            headers.setLocation(URI.create(encodePath(
-                    getBaseEnvLinkURL() + "/rest/openehr/v1/ehr/" + ehrId.toString() + "/directory/" + versionUid)));
+            headers.setLocation(createLocationUri(EHR, ehrId.toString(), DIRECTORY, versionUid));
             // TODO: Extract last modified from SysPeriod timestamp of fetched folder record
             headers.setLastModified(DateTime.now().getMillis());
         }
@@ -291,10 +289,8 @@ public class OpenehrDirectoryController extends BaseController implements Direct
             throw new PreconditionFailedException(
                     "If-Match version_uid does not match latest version.",
                     directoryId,
-                    encodePath(getBaseEnvLinkURL()
-                            + "/rest/openehr/v1/ehr/"
-                            + ehrId.toString()
-                            + "/directory/" + directoryId));
+                    createLocationUri(EHR, ehrId.toString(), DIRECTORY, directoryId)
+                            .toString());
         }
     }
 }