+++ service/src/main/java/org/ehrbase/dao/access/jooq/TenantAccess.java
@@ -17,17 +17,24 @@
  */
 package org.ehrbase.dao.access.jooq;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
+import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
+import org.apache.commons.lang3.StringUtils;
 import org.ehrbase.api.exception.InternalServerException;
 import org.ehrbase.api.tenant.Tenant;
 import org.ehrbase.dao.access.interfaces.I_TenantAccess;
+import org.ehrbase.functional.ExceptionalSupplier;
 import org.ehrbase.jooq.pg.Tables;
 import org.ehrbase.jooq.pg.tables.records.TenantRecord;
 import org.jooq.DSLContext;
+import org.jooq.JSON;
 import org.jooq.Result;
 
 public class TenantAccess implements I_TenantAccess {
@@ -45,6 +52,8 @@ public class TenantAccess implements I_TenantAccess {
         TenantRecord rec = ctx.newRecord(Tables.TENANT);
         rec.setTenantId(tenant.getTenantId());
         rec.setTenantName(tenant.getTenantName());
+        String json = mapToJson.apply(tenant.getTenantProperties());
+        rec.setTenantProperties(JSON.json(json));
         return rec;
     }
 
@@ -67,6 +76,22 @@ public class TenantAccess implements I_TenantAccess {
                 .orElse(null);
     }
 
+    private static Function<Map<String, Object>, String> mapToJson = map -> {
+        if (map == null) return null;
+
+        ExceptionalSupplier<String, Exception> sup = () -> new ObjectMapper().writeValueAsString(map);
+        return sup.get();
+    };
+
+    @SuppressWarnings("unchecked")
+    private static Function<JSON, Map<String, Object>> jsonToMap = json -> {
+        if (json == null || StringUtils.isEmpty(json.data())) return Collections.emptyMap();
+
+        ExceptionalSupplier<Map<String, Object>, Exception> sup =
+                () -> (Map<String, Object>) new ObjectMapper().readValue(json.data(), Map.class);
+        return sup.get();
+    };
+
     @Override
     public Tenant convert() {
         return new Tenant() {
@@ -76,6 +101,10 @@ public class TenantAccess implements I_TenantAccess {
 
             public String getTenantName() {
                 return record.getTenantName();
+            }
+
+            public Map<String, Object> getTenantProperties() {
+                return jsonToMap.apply(record.getTenantProperties());
             }
         };
     }
@@ -87,6 +116,8 @@ public class TenantAccess implements I_TenantAccess {
             new InternalServerException(String.format(ERR_TENANT_ID, tenant.getTenantId()));
 
         record.setTenantName(tenant.getTenantName());
+        String json = mapToJson.apply(tenant.getTenantProperties());
+        record.setTenantProperties(JSON.json(json));
         record.update();
         return convert();
     }